Project rename, Readme and contributing
Project was renamed from HeilHydra to Dawn. The readme was
restructured to be more of a general landing page, and
contributor-specific documentation was moved to a
separate document.

Some holes still remain to be filled; namely, link to the documentation
points to nowhere, and we are refering to some software which we are not
currently installing in this version of the codebase.