hyperv: removed diff. disk, virtal if, new image
The currently used hyperv image was having strange network
behaviours, so I switched to a more stable box.

On this image, setting up the main interface broke the network
restart (possibly because the connection got dropped since
we were changing its IP); instead, we now set a virtual interface
on the interface specified in instances.yml

Diffenrential disks appear to have an impact on perfromance, and
do not really make booting faster; removed the option.