if config override, vmswitch selection
The VMSwitch selection is now dynamic (instead of expecting the vmswitch
to always be in the same position).

We reconfigure the if instead of creating a virtual if because Hyper-V
DHCP doesn't seem to allow reconfiguration (meaning the network restart
would fail). We also restart the network asynchronously because since we
reconfigure the network, we would lose the connection otherwise.